+++ changes.md
@@ -46,3 +46,4 @@ and a vector of polynomials denoting constraints.
 - [ ] Verify no un-necessary files in the PR
 - [ ] performance benchmarking, does $C_{\alpha,j}$ get cast into dense formulation
 - [ ] a set of commuting and noncommuting variables
+- [ ] polynomial optimization problem on LJ Potential